Remove all uses of ilist_node::getNextNode() and ilist_node::getPrevNode() in favor of just using iterators.
There was churn upstream on ilist_node to remove undefined behavior. Rather than
updating the different ilist_node patches for the hacky change required to not
use iterators, just use iterators and keep everything as ilist_nodes. Upstream
they want to eventually do this, so it makes sense for us to just do it now.